Move EmailField to its own file
Testing prep
